fastapi_rpac/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.5

# 测试依赖
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.26.0
pytest-asyncio==0.23.5

# 开发工具
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 